[Refactor] Move column-to-arrow to Column
